pyproject.toml
[tool.poetry]
name = "pictoroute"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
geopy = "^2.4.1"
requests = "^2.32.3"
jupyter = "^1.1.1"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
pillow = "^11.0.0"
python-multipart = "^0.0.12"
pytesseract = "^0.3.13"
opencv-python = "^4.10.0.84"
black = "^24.10.0"
pydantic = "^2.9.2"
openai = "^1.52.0"
anthropic = "^0.37.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
